+++ src/Framework/ServiceProvider.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace App\Framework;
+
+abstract class ServiceProvider
+{
+    public function __construct(
+        protected Application $app
+    )
+    {
+
+    }
+
+    public abstract function register(): void;
+}