+++ tests/Framework/Fixtures/NeedsEntry.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Tests\Framework\Fixtures;
+
+class NeedsEntry
+{
+    public function __construct(
+        private Entry $entry
+    )
+    {
+
+    }
+}