+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Added `./bin/server` command as handy feature to run php development server
 - Added PHPUnit support
 - Added .env support
+- Added simple IOC Container
+- Added PDO Service